# Conflicts:
#	src/main/java/com/kh/magamGG/domain/agency/controller/AgencyController.java
#	src/main/java/com/kh/magamGG/domain/agency/dto/response/AgencyDetailResponse.java
#	src/main/java/com/kh/magamGG/domain/agency/entity/Agency.java
#	src/main/java/com/kh/magamGG/domain/agency/service/AgencyService.java
#	src/main/java/com/kh/magamGG/domain/agency/service/AgencyServiceImpl.java
#	src/main/java/com/kh/magamGG/domain/attendance/controller/AttendanceRequestController.java
#	src/main/java/com/kh/magamGG/domain/attendance/repository/AttendanceRequestRepository.java
#	src/main/java/com/kh/magamGG/domain/attendance/repository/LeaveBalanceRepository.java
#	src/main/java/com/kh/magamGG/domain/attendance/service/AttendanceService.java
#	src/main/java/com/kh/magamGG/domain/attendance/service/AttendanceServiceImpl.java
#	src/main/java/com/kh/magamGG/domain/member/controller/MemberController.java
#	src/main/java/com/kh/magamGG/domain/member/dto/response/MemberResponse.java
#	src/main/java/com/kh/magamGG/domain/member/service/MemberService.java
#	src/main/java/com/kh/magamGG/domain/member/service/MemberServiceImpl.java
#	src/main/resources/application.yaml
#	src/main/resources/schema.sql

+++ src/main/java/com/kh/magamGG/domain/attendance/entity/LeaveBalance.java
@@ -34,7 +34,7 @@ public class LeaveBalance {
 	private Integer leaveBalanceUsedDays;
 	
 	@Column(name = "LEAVE_BALANCE_REMAIN_DAYS")
-	private Integer leaveBalanceRemainDays;
+	private Double leaveBalanceRemainDays;
 	
 	@Column(name = "LEAVE_BALANCE_YEAR", nullable = false, length = 4)
 	private String leaveBalanceYear;
